requirements.txt
Django
requests
whitenoise
wheel
openai
azure-cognitiveservices-speech
